backend/requirements.txt
# ══════════════════════════════════════════════
# SaiSuite Backend — Dependencias Python
# ══════════════════════════════════════════════

# Django core
Django==6.0.3
djangorestframework==3.17.0
django-environ==0.11.2
django-cors-headers==4.4.0
django-filter==24.3

# Auth
djangorestframework-simplejwt==5.3.1

# Base de datos
psycopg2-binary==2.9.9

# AWS
boto3==1.34.144
botocore==1.34.144

# IA
openai==1.35.14

# Utilidades
Pillow==10.4.0
python-dateutil==2.9.0
pytz==2024.1
openpyxl==3.1.5

# Dev / Testing
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==25.8.0

# Calidad de código
black==24.4.2
flake8==7.1.0
isort==5.13.2
